FR-702: Move create invoice template module to right place. Fix tests
