+++ src/app/sections/payment-section/integrations/payment-link/payment-link.module.ts
@@ -19,8 +19,8 @@ import { FormControlsModule } from '@dsh/components/form-controls';
 import { LayoutModule } from '@dsh/components/layout';
 import { ConfirmActionDialogModule } from '@dsh/components/popups';
 
-import { CreateInvoiceTemplateModule } from '../../../create-invoice-template';
 import { CreateInvoiceOrInvoiceTemplateComponent } from './create-invoice-or-invoice-template';
+import { CreateInvoiceTemplateModule } from './create-invoice-template';
 import { PaymentLinkRoutingModule } from './payment-link-routing.module';
 import { PaymentLinkComponent } from './payment-link.component';
