+++ src/app/sections/payment-section/integrations/shops/shop-creation/components/create-shop-dialog/create-shop-dialog.component.spec.ts
@@ -40,7 +40,7 @@ describe('CreateShopDialogComponent', () => {
                 }),
                 RouterTestingModule.withRoutes([
                     {
-                        path: 'onboarding',
+                        path: 'claim-section/onboarding',
                         component: MockOnBoardingComponent,
                     },
                 ]),
